Move lazyClientHost to own file
