+++ extensions/typescript-language-features/src/extension.ts
@@ -9,16 +9,10 @@ import { Api, getExtensionApi } from './api';
 import { registerCommands } from './commands/index';
 import { LanguageConfigurationManager } from './features/languageConfiguration';
 import * as task from './features/task';
-import TypeScriptServiceClientHost from './typeScriptServiceClientHost';
-import { flatten } from './utils/arrays';
+import { createLazyClientHost, lazilyActivateClient } from './lazyClientHost';
 import { CommandManager } from './utils/commandManager';
 import * as electron from './utils/electron';
-import * as fileSchemes from './utils/fileSchemes';
-import { standardLanguageDescriptions } from './utils/languageDescription';
-import * as ProjectStatus from './utils/largeProjectStatus';
-import { lazy, Lazy } from './utils/lazy';
 import LogDirectoryProvider from './utils/logDirectoryProvider';
-import ManagedFileContextManager from './utils/managedFileContext';
 import { PluginManager } from './utils/plugins';
 
 export function activate(
@@ -33,7 +27,9 @@ export function activate(
 	const onCompletionAccepted = new vscode.EventEmitter<vscode.CompletionItem>();
 	context.subscriptions.push(onCompletionAccepted);
 
-	const lazyClientHost = createLazyClientHost(context, pluginManager, commandManager, item => {
+	const logDirectoryProvider = new LogDirectoryProvider(context);
+
+	const lazyClientHost = createLazyClientHost(context, pluginManager, commandManager, logDirectoryProvider, item => {
 		onCompletionAccepted.fire(item);
 	});
 
@@ -48,84 +44,6 @@ export function activate(
 	context.subscriptions.push(lazilyActivateClient(lazyClientHost, pluginManager));
 
 	return getExtensionApi(onCompletionAccepted.event, pluginManager);
-}
-
-function createLazyClientHost(
-	context: vscode.ExtensionContext,
-	pluginManager: PluginManager,
-	commandManager: CommandManager,
-	onCompletionAccepted: (item: vscode.CompletionItem) => void,
-): Lazy<TypeScriptServiceClientHost> {
-	return lazy(() => {
-		const logDirectoryProvider = new LogDirectoryProvider(context);
-
-		const clientHost = new TypeScriptServiceClientHost(
-			standardLanguageDescriptions,
-			context.workspaceState,
-			pluginManager,
-			commandManager,
-			logDirectoryProvider,
-			onCompletionAccepted);
-
-		context.subscriptions.push(clientHost);
-
-		clientHost.serviceClient.onReady(() => {
-			context.subscriptions.push(
-				ProjectStatus.create(
-					clientHost.serviceClient,
-					clientHost.serviceClient.telemetryReporter));
-		});
-
-		return clientHost;
-	});
-}
-
-function lazilyActivateClient(
-	lazyClientHost: Lazy<TypeScriptServiceClientHost>,
-	pluginManager: PluginManager,
-) {
-	const disposables: vscode.Disposable[] = [];
-
-	const supportedLanguage = flatten([
-		...standardLanguageDescriptions.map(x => x.modeIds),
-		...pluginManager.plugins.map(x => x.languages)
-	]);
-
-	let hasActivated = false;
-	const maybeActivate = (textDocument: vscode.TextDocument): boolean => {
-		if (!hasActivated && isSupportedDocument(supportedLanguage, textDocument)) {
-			hasActivated = true;
-			// Force activation
-			void lazyClientHost.value;
-
-			disposables.push(new ManagedFileContextManager(resource => {
-				return lazyClientHost.value.serviceClient.toPath(resource);
-			}));
-			return true;
-		}
-		return false;
-	};
-
-	const didActivate = vscode.workspace.textDocuments.some(maybeActivate);
-	if (!didActivate) {
-		const openListener = vscode.workspace.onDidOpenTextDocument(doc => {
-			if (maybeActivate(doc)) {
-				openListener.dispose();
-			}
-		}, undefined, disposables);
-	}
-
-	return vscode.Disposable.from(...disposables);
-}
-
-function isSupportedDocument(
-	supportedLanguage: string[],
-	document: vscode.TextDocument
-): boolean {
-	if (supportedLanguage.indexOf(document.languageId) < 0) {
-		return false;
-	}
-	return fileSchemes.isSupportedScheme(document.uri.scheme);
 }
 
 export function deactivate() {
